worked on the DB migrations and AddPriceProposal logic in the PriceNegotiationService

+++ Services/PriceNegotiationService.cs
@@ -12,9 +12,9 @@ public class PriceNegotiationService : IPriceNegotiationService
     
     private readonly IMapper _mapper;
     private readonly PriceNegotiationDbContext _dbContext;
-    private readonly ILogger<ProductCatalogueService> _logger; //check if needed
+    private readonly ILogger<PriceNegotiationService> _logger; //check if needed
 
-    public PriceNegotiationService(IMapper mapper, PriceNegotiationDbContext dbContext, ILogger<ProductCatalogueService> logger)
+    public PriceNegotiationService(IMapper mapper, PriceNegotiationDbContext dbContext, ILogger<PriceNegotiationService> logger)
     {
         _mapper = mapper;
         _dbContext = dbContext;
@@ -29,36 +29,30 @@ public class PriceNegotiationService : IPriceNegotiationService
 
     }
 
-    public void AddPriceProposal(PriceProposalDto priceProposal) 
+    public PriceProposalDto AddPriceProposal(PriceProposalDto priceProposal) 
     {
         var proposal =  _mapper.Map<PriceProposal>(priceProposal);
         _dbContext.PriceProposals.Add(proposal);
 
-        if (proposal.Accepted == false & proposal.AttemptsLeft>=0)
+        if (proposal.Accepted == false && proposal.AttemptsLeft>=0)
         {
-            if (priceProposal.ProposedPrice1 == 2 * (proposal.ProductPrice))
+            if (priceProposal.ProposedPrice == 2 * (proposal.ProductPrice))
             {
-                throw new Exception("The proposed price is double the product price. The price proposal is rejected.");
                 _dbContext.PriceProposals.Remove(proposal);
+                throw new Exception("The proposed price is double the product price. The price proposal is rejected.");
             }
-            if (priceProposal.ProposedPrice1 > proposal.ProductPrice)
-            {
-                _dbContext.SaveChanges();
-            }
-            else if (priceProposal.ProposedPrice1 < proposal.ProductPrice)
+            if (priceProposal.ProposedPrice < proposal.ProductPrice)
             {
                 throw new BadRequestException("The proposed price is lower than the product price.");
             }
+            if (priceProposal.ProposedPrice > proposal.ProductPrice)
+            {
+                _dbContext.SaveChanges();
+            }
+            proposal.AttemptsLeft--;
         }
-        else if (proposal.Accepted == true)
-        {
-            throw new Exception("The price proposal has already been accepted.");
-        }
-        else if (proposal.AttemptsLeft <= 0)
-        {
-            throw new Exception("The price proposal on this item is unavailable.");
-        }
-        proposal.AttemptsLeft--;
+        
+        return _mapper.Map<PriceProposalDto>(proposal);
     }
 
 }
